+++ lib/elixir_script/translator/utils.ex
@@ -177,6 +177,14 @@ defmodule ElixirScript.Translator.Utils do
     "__in__"
   end
 
+  def filter_name(:new) do
+    "__new__"
+  end
+
+  def filter_name(:delete) do
+    "__delete__"
+  end
+
   def filter_name(name) do
     to_string(name)
     |> String.replace("?", "__qmark__")