+++ priv/javascript/elixir.js
@@ -20,4 +20,6 @@ export { default as Bitwise } from './lib/bitwise';
 export { default as Enumerable } from './lib/enumerable';
 export { default as Collectable } from './lib/collectable';
 export { default as Inspect } from './lib/inspect';
+export { default as Map } from './lib/map';
+export { default as MapSet } from './lib/map_set';